Update CalculateGroupPath.R
In present form the levels information of the groups is lost when the group path radial x-y coordinates are calculated so ggplot always plots groups in alphabetical or numerical order. Added line 41 in CalculateGroupPath.R creates a factor with the same levels as original dataframe so initial group order is not lost.